eliminated redundant code

+++ Calc.js
@@ -9,57 +9,31 @@ app.controller('CalcController', ['$scope', '$window', function ($scope, $window
 	$scope.input1=null;
 	$scope.input2=null;
 	
-		$scope.addNum = function(){
+		$scope.operation = function(operator){
 			if ($scope.input1 == null || $scope.input2 == null){
 				alert("Both operands are required");
 				return;
-			}else{
-			$scope.result = $scope.input1 + $scope.input2;
-			};
-		$scope.para = "The result of adding " + $scope.input1 + " and " + $scope.input2 + " is " + $scope.result;
-		$scope.input1 = 0;
-		$scope.input2 = 0;
-		}
-		
-		$scope.subtractNum = function(){
-			if ($scope.input1 == null || $scope.input2 == null){
-				alert("Both operands are required");
-				return;
-			}else{
-			$scope.result = $scope.input1 - $scope.input2;
-			};
-		$scope.para = "The result of subtracting " + $scope.input1 + " and " + $scope.input2 + " is " + $scope.result;
-		$scope.input1 = 0;
-		$scope.input2 = 0;
-		}	
-	
-		$scope.multiplyNum = function(){
-			if ($scope.input1 == null || $scope.input2 == null){
-				alert("Both operands are required");
-				return;
-			}else{
-			$scope.result = $scope.input1 * $scope.input2;
-			};
-		$scope.para = "The result of multiplying " + $scope.input1 + " and " + $scope.input2 + " is " + $scope.result;
-		$scope.input1 = 0;
-		$scope.input2 = 0;
-		}
-		
-		$scope.divideNum = function(){
-			if ($scope.input1 == null || $scope.input2 == null){
-				alert("Both operands are required");
-				return;
-			}else if ($scope.input2 == 0){
+			}else if(operator == "+"){
+				$scope.result = $scope.input1 + $scope.input2;
+				$scope.para = "The result of adding " + $scope.input1 + " and " + $scope.input2 + " is " + $scope.result;
+			}else if(operator == "-"){
+				$scope.result = $scope.input1 - $scope.input2;
+				$scope.para = "The result of subtracting " + $scope.input1 + " and " + $scope.input2 + " is " + $scope.result;
+			}else if(operator == "*"){
+				$scope.result = $scope.input1 * $scope.input2;
+				$scope.para = "The result of multiplying " + $scope.input1 + " and " + $scope.input2 + " is " + $scope.result;
+			}else if(operator == "/"){
+				if($scope.input2 == 0){
 				alert("Are you trying to kill us all? Please don't divide by zero");
 				return;
-			}else{
-			$scope.result = $scope.input1 / $scope.input2;
-			};
-		$scope.para = "The result of dividing " + $scope.input1 + " and " + $scope.input2 + " is " + $scope.result;
-		$scope.input1 = 0;
-		$scope.input2 = 0;
-		}
+				}else{
+				$scope.result = $scope.input1 / $scope.input2;
+				$scope.para = "The result of dividing " + $scope.input1 + " and " + $scope.input2 + " is " + $scope.result;
+				};
+			}
+			$scope.input1 = 0;
+			$scope.input2 = 0;	
+		};
 		
-
 }]);
 })();
